get_next_line

+++ Common_Core/get_next_line/get_next_line_utils_bonus.c
@@ -6,7 +6,7 @@
 /*   By: thda-sil <thda-sil@student.42.rio>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/01/29 14:57:11 by thda-sil          #+#    #+#             */
-/*   Updated: 2024/01/29 15:59:39 by thda-sil         ###   ########.fr       */
+/*   Updated: 2024/01/29 19:00:57 by thda-sil         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -40,7 +40,7 @@ size_t	ft_strlen(const char *str)
 char	*ft_strjoin(char *s1, char *s2)
 {
 	char	*str;
-	int		i;
+	size_t	i;
 
 	i = 0;
 	str = (char *) malloc((ft_strlen(s1) + ft_strlen(s2) + 1) * sizeof(char));
